crawler/pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "playwright (>=1.55.0,<2.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "pylint (>=3.3.8,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-ollama (>=1.0.0,<2.0.0)",
    "pydantic (>=2.12.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "crawler", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
